(#147) Moved files in order to simplify imports
