+++ data/GameObjects/Wall.py
@@ -1,15 +1,12 @@
 import pygame
-from Color import invert_color
+from ..editor import *
 
-class Plate(pygame.sprite.Sprite):
-    def __init__(self, x,y,w,h, color = (255,0,0), linked_to_id = -1):
+class Wall(pygame.sprite.Sprite):
+    def __init__(self, x,y,w,h, color = (255,0,0)):
         super().__init__()
-        self.rect = pygame.Rect(x,y,w,30)
+        self.rect = pygame.Rect(x,y,w,h)
         self.color = color
-        self.collide = False
-        self.linked_to_id = linked_to_id
-        self.linked_to = None
-        self.players_on = []
+        self.collide = True
         self.image = pygame.Surface((w,h))
         self.image.fill(self.color)
 
@@ -29,20 +26,15 @@ class Plate(pygame.sprite.Sprite):
         return self.collide
 
     def get_properties(self):
-        return ["Linker"]
+        return ["ColorPicker"]
 
     def on_collision(self, collider):
-        if not self.players_on: #s'il n'y a aucun joueur sur la plaque
-            self.linked_to.switch_status()
-        if not collider in self.players_on:
-            self.players_on.append(collider)
+        return
 
     def on_collision_exit(self, collider):
-        print('exit')
-        self.players_on.remove(collider)
-        if not self.players_on:
-            self.linked_to.switch_status()
+        return
 
     def as_string(self):
+        color_int = (int(v) for v in self.color)
         rect_int = [ int(self.rect.x), int(self.rect.y), int(self.rect.w), int(self.rect.h) ]
-        return 'Plate, {},{},{},{}, {}\n'.format(*rect_int, self.linked_to_id)
+        return 'Wall, {},{},{},{}, {},{},{}\n'.format(*rect_int, *color_int)
