+++ data/GameObjects/__init__.py
@@ -1,0 +1,5 @@
+from .Wall import Wall
+from .Door import Door
+from .Plate import Plate
+from .SpawnPoint import SpawnPoint
+from .Player import Player
