+++ data/editor/__init__.py
@@ -1,0 +1,4 @@
+from .Camera import Camera
+from .Color import *
+from .ColorPicker import ColorPicker
+from .PropertyPanel import PropertyPanel
